+++ models/role.mjs
@@ -25,6 +25,7 @@ export const ROLE_ANONYMOUS = new RoleList(
   rights.CAN_GET_RIDE_POSITION,
   rights.CAN_LIST_CAMPUS_BASIC,
   rights.CAN_SEND_CREATION_TOKEN,
+  rights.CAN_CREATE_RATING,
 );
 roles.ROLE_ANONYMOUS = ROLE_ANONYMOUS;
 
@@ -192,6 +193,8 @@ export const ROLE_SUPERADMIN = new RoleList(
   rights.CAN_ADD_ROLE_DRIVER,
 
   rights.CAN_LIST_ALL_CAMPUSES,
+
+  rights.CAN_LIST_RATING,
 );
 roles.ROLE_SUPERADMIN = ROLE_SUPERADMIN;