+++ routes/users.mjs
@@ -1,5 +1,6 @@
 import generateCRUD from '../helpers/abstract-route';
 import User from '../models/user';
+import NotificationDevice from '../models/notification-device';
 import {
   CAN_CREATE_USER, CAN_EDIT_SELF_USER_NAME, CAN_EDIT_SELF_USER_PASSWORD,
   CAN_SEND_CREATION_TOKEN,
@@ -7,7 +8,8 @@ import {
   CAN_EDIT_USER_SENSITIVE_DATA,
   CAN_GET_USER,
   CAN_LIST_USER,
-  CAN_REMOVE_USER,
+  CAN_REMOVE_USER, CAN_REMOVE_SELF_USER,
+  CAN_EDIT_USER_WITHOUT_UPPER_RIGHTS,
 } from '../models/rights';
 import config from '../services/config';
 
@@ -92,7 +94,8 @@ const router = generateCRUD(User, {
     lean: false,
   },
   delete: {
-    right: CAN_REMOVE_USER,
+    paramId: 'user_id',
+    right: [CAN_REMOVE_USER, CAN_REMOVE_SELF_USER],
   },
   list: {
     right: CAN_LIST_USER,
@@ -181,7 +184,11 @@ const router = generateCRUD(User, {
       if (!user.email_confirmed && body.email && body.email_token) {
         await user.confirmEmail(body.email_token);
       }
-
+      ctx.assert(
+        ctx.may(CAN_EDIT_USER_WITHOUT_UPPER_RIGHTS, user),
+        403,
+        'You\'re not authorized to update this user',
+      );
       ctx.body = await user.save();
 
       if ((ctx.headers[X_SEND_TOKEN] && ctx.headers[X_SEND_TOKEN] !== 'false')) {
@@ -214,5 +221,17 @@ const router = generateCRUD(User, {
     },
   },
 });
+
+router.post(
+  '/:userId/subscribe-device',
+  async (ctx) => {
+    const { request: { body } } = ctx;
+    body.user = {
+      _id: ctx.params.userId,
+    };
+    await NotificationDevice.findOneAndUpdateByUser(body);
+    ctx.status = 204;
+  },
+);
 
 export default router.routes();